requirements.txt
pynput
Pillow
pyperclip
psutil
